feat(product): add exception manage

+++ src/products/services/products.service.ts
@@ -1,4 +1,4 @@
-import { Injectable } from '@nestjs/common';
+import { HttpException, HttpStatus, Injectable } from "@nestjs/common";
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { Product } from '../entity/product.entity';
@@ -12,25 +12,55 @@ export class ProductsService {
     private readonly productRepository: Repository<Product>,
   ) {}
 
-  findAll(): Promise<Product[]> {
-    return this.productRepository.find();
+  async findAll(): Promise<Product[]> {
+    try {
+      return await this.productRepository.find();
+    } catch (error) {
+      throw new HttpException('Error fetching products', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
-  findOne(id: number): Promise<Product> {
-    return this.productRepository.findOneBy({ id });
+  async findOne(id: number): Promise<Product> {
+    const product = await this.productRepository.findOneBy({ id });
+    if (!product) {
+      throw new HttpException('Product not found', HttpStatus.NOT_FOUND);
+    }
+    return product;
   }
 
   async create(createProductDto: CreateProductDto): Promise<Product> {
-    const product = this.productRepository.create(createProductDto);
-    return this.productRepository.save(product);
+    try {
+      const product = this.productRepository.create(createProductDto);
+      return await this.productRepository.save(product);
+    } catch (error) {
+      throw new HttpException('Error creating product', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
   async update(id: number, updateProductDto: UpdateProductDto): Promise<Product> {
-    await this.productRepository.update(id, updateProductDto);
-    return this.productRepository.findOneBy({ id });
+    const product = await this.productRepository.findOneBy({ id });
+    if (!product) {
+      throw new HttpException('Product not found', HttpStatus.NOT_FOUND);
+    }
+
+    try {
+      await this.productRepository.update(id, updateProductDto);
+      return await this.productRepository.findOneBy({ id });
+    } catch (error) {
+      throw new HttpException('Error updating product', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 
   async delete(id: number): Promise<void> {
-    await this.productRepository.delete(id);
+    const product = await this.productRepository.findOneBy({ id });
+    if (!product) {
+      throw new HttpException('Product not found', HttpStatus.NOT_FOUND);
+    }
+
+    try {
+      await this.productRepository.delete(id);
+    } catch (error) {
+      throw new HttpException('Error deleting product', HttpStatus.INTERNAL_SERVER_ERROR);
+    }
   }
 }
